2nd Place/requirements.txt
albumentations
cv2
jpeg4py
matplotlib
numpy
pandas
Pillow
pytorch
pystac
rasterio
segmentation-models-pytorch
sklearn
tqdm
